requirements.txt
torch==2.0.1
transformers==4.35.0
sentence-transformers==2.2.2
langchain==0.0.340
faiss-cpu==1.7.4
pypdf2==3.0.1
requests==2.31.0
python-dotenv==1.0.1